+++ src/main/java/com/limited/sales/exception/sub/BadRequestException.java
@@ -1,10 +1,6 @@
 package com.limited.sales.exception.sub;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class BadRequestException extends RuntimeException {
+public final class BadRequestException extends RuntimeException {
   public BadRequestException(String message) {
     super(message);
   }
