+++ src/main/java/com/limited/sales/exception/sub/DuplicatedIdException.java
@@ -1,10 +1,6 @@
 package com.limited.sales.exception.sub;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class DuplicatedIdException extends RuntimeException {
+public final class DuplicatedIdException extends RuntimeException {
   public DuplicatedIdException(String message) {
     super(message);
   }
